add `media/ui` tests
